refactor(bayn): modularize evidence recovery
